Re-introduce ServiceGraphVertex. Convert allvertex classes to AutoValue
