+++ graph-test/src/test/java/net/spals/appbuilder/graph/model/PrintableVertexTest.java
@@ -7,8 +7,10 @@ import org.mockito.Mockito;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
-import static net.spals.appbuilder.graph.model.ServiceDAGVertex.createVertex;
-import static net.spals.appbuilder.graph.model.ServiceDAGVertex.createVertexWithProvider;
+import static net.spals.appbuilder.graph.model.PrintableVertex.createPrintableVertex;
+import static net.spals.appbuilder.graph.model.ServiceDAGVertex.createDAGVertex;
+import static net.spals.appbuilder.graph.model.ServiceDAGVertex.createDAGVertexWithProvider;
+import static net.spals.appbuilder.graph.model.ServiceGraphVertex.createGraphVertex;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.mockito.Mockito.verify;
@@ -22,8 +24,8 @@ public class PrintableVertexTest {
 
     @Test
     public void tetGetGuiceKey() {
-        final ServiceDAGVertex vertex = Mockito.mock(ServiceDAGVertex.class);
-        final PrintableVertex printableVertex = new PrintableVertex(vertex, " ");
+        final IServiceGraphVertex vertex = Mockito.mock(IServiceGraphVertex.class);
+        final PrintableVertex printableVertex = createPrintableVertex(vertex, " ");
 
         printableVertex.getGuiceKey();
         verify(vertex).getGuiceKey();
@@ -31,30 +33,21 @@ public class PrintableVertexTest {
 
     @Test
     public void tetGetServiceInstance() {
-        final ServiceDAGVertex vertex = Mockito.mock(ServiceDAGVertex.class);
-        final PrintableVertex printableVertex = new PrintableVertex(vertex, " ");
+        final IServiceGraphVertex vertex = Mockito.mock(IServiceGraphVertex.class);
+        final PrintableVertex printableVertex = createPrintableVertex(vertex, " ");
 
         printableVertex.getServiceInstance();
         verify(vertex).getServiceInstance();
     }
 
-    @Test
-    public void tetGetProviderSource() {
-        final ServiceDAGVertex vertex = Mockito.mock(ServiceDAGVertex.class);
-        final PrintableVertex printableVertex = new PrintableVertex(vertex, " ");
-
-        printableVertex.getProviderSource();
-        verify(vertex).getProviderSource();
-    }
-
     @DataProvider
     Object[][] toStringProvider() {
         return new Object[][] {
-            {createVertex(Key.get(String.class), "1"), " ", "\"1\""},
-            {createVertex(Key.get(String.class, Names.named("constant")), "1"), "~",
+            {createGraphVertex(Key.get(String.class), "1"), " ", "\"1\""},
+            {createGraphVertex(Key.get(String.class, Names.named("constant")), "1"), "~",
                 "@Named(constant)~\"1\""},
-            {createVertexWithProvider(createVertex(Key.get(PrintableVertexTest.class), new PrintableVertexTest()),
-                createVertex(Key.get(Provider.class), new Provider<PrintableVertexTest>() {
+            {createDAGVertexWithProvider(Key.get(PrintableVertexTest.class), new PrintableVertexTest(),
+                createDAGVertex(Key.get(Provider.class), new Provider<PrintableVertexTest>() {
                     @Override
                     public PrintableVertexTest get() {
                         return new PrintableVertexTest();
@@ -66,10 +59,10 @@ public class PrintableVertexTest {
     }
 
     @Test(dataProvider = "toStringProvider")
-    public void testToString(final ServiceDAGVertex<?> vertex,
+    public void testToString(final IServiceGraphVertex<?> vertex,
                              final String separator,
                              final String expectedResult) {
-        final PrintableVertex printableVertex = new PrintableVertex(vertex, separator);
+        final PrintableVertex printableVertex = createPrintableVertex(vertex, separator);
         assertThat(printableVertex.toString(), is(expectedResult));
     }
 }
